+++ app/models/legal_services/rm6240/admin/supplier_contact_detail.rb
@@ -4,7 +4,7 @@ module LegalServices
       class SupplierContactDetail < ::Supplier::Framework::ContactDetail
         self.table_name = 'supplier_framework_contact_details'
 
-        ATTRIBUTES_TO_SKIP_VALIDATION = %i[name lot_4a_prospectus_link lot_4b_prospectus_link lot_4c_prospectus_link lot_5_prospectus_link].freeze
+        ATTRIBUTES_TO_SKIP_VALIDATION = %i[name lot_4a_prospectus_link lot_4b_prospectus_link lot_4c_prospectus_link lot_5_prospectus_link managed_service_provider_name managed_service_provider_telephone managed_service_provider_email].freeze
       end
     end
   end